requirements.txt
flask
mysql-connector-python
pandas
matplotlib
